// frontend/gamejam-tickets/src/app/models/suppChatTicket.model.ts
export interface SuppChatTicket {
  _id: string;               
  idUserIssued: string;      
  userName: string;          
  idSupport: string;         
  supportName: string;
  idCategory: string;         
  categoryName: string;       
  resolutionState: string;   
  closureState: string;      
  topic: string;             
}